+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [#127](https://github.com/deviceinsight/kafkactl/issues/127) the commands `create consumer-group` and `reset consumer-group-offset` can now be called with multiple `--topic` parameters.
+  Additionally, `reset consumer-group-offset` can be called with `--all-topics` parameter, which will reset all topics in the group.
+
 ## 2.2.1 - 2022-04-04
 
 ### Fixed